Added logic to handle drag selections
Drag selections are now handled better using a workaround way.
The plugin command enters a "standby" mode instead of an "inactive" mode
and then listens for selection changes that are within 1 move difference.
(Happens only if the new selection is single selection and not "multisel"... behaviour then is hard to define. Sadly, this means column-select will not be
handled properly)

Work left on this feature:
 - add a setting to enable/disable this
 - add documentation