Fixed buggy behaviour of 'combine-1move' feature
The 'combine-onemove-events' feature was buggy because the variable that
stored LAST_SELECTION region, was incorrectly set to the first region of
the current selection (which surprisingly included the STORED_SELECTION
regions). Now, the selection set is trimmed of the STORED_SELECTION and
only then is the first region taken (should always be exactly one).