Added a timeout option to delay actual completion
Added a setting for a feature that would delay the actual completion of
the command by a specified time (in miliseconds). This will allow for
multiple changes to the selection to be handled by the command.

THe feature is potentially very confusing, so by default it is disabled.
This is effectively achieved by setting the value to zero or a negative
number.