Updated Keybindings files
Updated the Default keybindings file. The platform dependent versions were
removed - primarily because I have no way of properly testing if they are
correct, and secondary, they most likely would be the same as the Default
file anyway.

Added the keybindings for use of the command in "manual completion" mode.
That includes a keybinding (ctrl+alt+m) for triggering the command and a
keybinding (ctrl+alt+m again) to finish / complete the command.