Fixed errors in the 'manual completion' behaviour
+ fixed a mistake where part of the logic was not put into the code. That
missing code that initialises the PENDING_TIMEOUTS variable is now there.

+ fixed an error where the STORED_SELECTION was not updated. This meant
that only the initial selection was actually stored and the subsequent
selection changes overrode each other rather than adding to each other.

+ fixed an error where the manual "complete" call would not run. This was
cause by sublime (correctly) believing the command was unavailable. The
accessor was changed to always say the command is available. Sadly, there
is no way (that I know of) to decide availability based on parameters
passed to command - all calls are treated the same way.